+++ tests/tests.sh
@@ -163,37 +163,37 @@ testUrlDefaultNameTrailingSlash()
     assertContains "Verify whether 'wcurl' chooses the correct default filename when there's no path in the URL and the URl ends with a slash" "${ret}" 'index.html'
 }
 
-testUrlDecodingWhitespaces()
+testUrlDecodingWhitespace()
 {
     url='example.com/filename%20with%20spaces'
     ret=$(${WCURL_CMD} ${url} 2>&1)
-    assertContains "Verify whether 'wcurl' successfully decodes percent-encoded whitespaces in URLs" "${ret}" 'filename with spaces'
+    assertContains "Verify whether 'wcurl' successfully decodes percent-encoded whitespace in URLs" "${ret}" 'filename with spaces'
 }
 
-testUrlDecodingWhitespacesTwoFiles()
+testUrlDecodingWhitespaceTwoFiles()
 {
     url='example.com/filename%20with%20spaces'
     url_2='example.com/filename2%20with%20spaces'
     ret=$(${WCURL_CMD} ${url} ${url_2} 2>&1)
-    assertContains "Verify whether 'wcurl' successfully decodes percent-encoded whitespaces in URLs" "${ret}" 'filename with spaces'
-    assertContains "Verify whether 'wcurl' successfully decodes percent-encoded whitespaces in URLs" "${ret}" 'filename2 with spaces'
+    assertContains "Verify whether 'wcurl' successfully decodes percent-encoded whitespace in URLs" "${ret}" 'filename with spaces'
+    assertContains "Verify whether 'wcurl' successfully decodes percent-encoded whitespace in URLs" "${ret}" 'filename2 with spaces'
 }
 
 testUrlDecodingDisabled()
 {
     url='example.com/filename%20with%20spaces'
     ret=$(${WCURL_CMD} --no-decode-filename ${url} 2>&1)
-    assertContains "Verify whether 'wcurl' successfully decodes percent-encoded whitespaces in URLs" "${ret}" 'filename%20with%20spaces'
+    assertContains "Verify whether 'wcurl' successfully decodes percent-encoded whitespace in URLs" "${ret}" 'filename%20with%20spaces'
 }
 
-testUrlDecodingWhitespacesQueryString()
+testUrlDecodingWhitespaceQueryString()
 {
     url='example.com/filename%20with%20spaces?query=string'
     ret=$(${WCURL_CMD} "${url}" 2>&1)
-    assertContains "Verify whether 'wcurl' successfully decodes percent-encoded whitespaces in URLs with query strings" "${ret}" 'filename with spaces'
+    assertContains "Verify whether 'wcurl' successfully decodes percent-encoded whitespace in URLs with query strings" "${ret}" 'filename with spaces'
 }
 
-testUrlDecodingWhitespacesTrailingSlash()
+testUrlDecodingWhitespaceTrailingSlash()
 {
     url='example.com/filename%20with%20spaces/'
     ret=$(${WCURL_CMD} ${url} 2>&1)